Handle mul instructions with two immediates on x86-32 and x86-64
We were hitting an assert in the assembler due to an instruction that
had not been properly legalized. Now we ensure that mul instructions
take at most one immediate.

It is not clear why these instructions were not removed by upstream
optimizations. This is possibly a result of values only being known at
link time but somehow missed by LTO.

This CL includes some minor cleanup as well:
* Running lit tests does not actually require the Subzero runtime, so
  that is removed from the dependencies list.
* Removes unused capture clauses that were causing warnings treated as
  errors on newer compilers.

Change-Id: I8dc561c82e309fedb46475dcacf14dcb7b535b99
Reviewed-on: https://chromium-review.googlesource.com/896962
Reviewed-by: Jim Stichnoth <stichnot@chromium.org>
Reviewed-by: Karl Schimpf <kschimpf@chromium.org>